feat: voting pages and events

+++ server/src/polls/polls.service.ts
@@ -102,7 +102,7 @@ export class PollsService {
   ): Promise<Poll | void> {
     const poll = await this.pollsRepository.getPoll(pollID);
 
-    if (!poll.hasStarted) {
+    if (poll && !poll.hasStarted) {
       const updatedPoll = await this.pollsRepository.removeParticipant(
         pollID,
         userID,
